feat: delete cloudinary images on motor delete

+++ app/api/admin/repas/[motorId]/route.ts
@@ -1,6 +1,26 @@
 import { NextResponse } from "next/server";
 
 import prisma from "@/app/libs/prismadb";
+import cloudinary from "cloudinary";
+
+// Configure Cloudinary (ensure that your Cloudinary credentials are set in environment variables)
+cloudinary.v2.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+});
+
+// Helper function to extract public_id from Cloudinary URL
+const extractPublicId = (url: string) => {
+  const parts = url.split("/");
+  const versionIndex = parts.findIndex((part) => part.startsWith("v"));
+  const publicIdWithExt = parts.slice(versionIndex + 1).join("/");
+  const publicId = publicIdWithExt.substring(
+    0,
+    publicIdWithExt.lastIndexOf(".")
+  );
+  return publicId;
+};
 
 export async function DELETE(
   req: Request,
@@ -13,7 +33,32 @@ export async function DELETE(
       return new NextResponse("Missing fields, id", { status: 400 });
     }
 
-    const motor = await prisma.motor.delete({
+    // Fetch the images of the motor from the database
+    const motor = await prisma.motor.findUnique({
+      where: {
+        id: motorId,
+      },
+      select: {
+        images: true,
+      },
+    });
+
+    if (!motor) {
+      return new NextResponse("Motor not found", { status: 404 });
+    }
+
+    const { images } = motor;
+
+    // Delete each image from Cloudinary
+    const deletePromises = images.map((imageUrl: string) => {
+      const publicId = extractPublicId(imageUrl);
+      return cloudinary.v2.api.delete_resources([publicId]);
+    });
+
+    // Wait for all Cloudinary deletions to complete
+    await Promise.all(deletePromises);
+
+    await prisma.motor.delete({
       where: {
         id: motorId,
       },
